+++ CHANGES.md
@@ -40,6 +40,8 @@ Version 0.20.0
 - Changes to api `onFilesAdded`
 - Multiple file uploads added as option to `uploader` attribute
 - `position` method signature changed to `position(top, left)`
+- Remove `ListMethod` methods
+- Remove `list` method
 
 Version 0.10.0
 -----------